+++ lib/src/exceptions.dart
@@ -1,5 +1,6 @@
 import 'events.dart';
 import 'message.dart';
+import 'push.dart';
 import 'socket.dart';
 
 /// Exception yield when a [PhoenixSocket] closes for unexpected reasons.
@@ -12,19 +13,19 @@ class PhoenixException implements Exception {
   });
 
   /// The associated error event.
-  final PhoenixSocketErrorEvent socketError;
+  final PhoenixSocketErrorEvent? socketError;
 
   /// The associated close event.
-  final PhoenixSocketCloseEvent socketClosed;
+  final PhoenixSocketCloseEvent? socketClosed;
 
   /// The associated channel event.
-  final String channelEvent;
+  final String? channelEvent;
 
   /// The error message for this exception.
-  Message get message {
-    if (socketClosed is PhoenixSocketCloseEvent) {
+  Message? get message {
+    if (socketClosed != null) {
       return Message(event: PhoenixChannelEvent.error);
-    } else if (socketError is PhoenixSocketErrorEvent) {
+    } else if (socketError != null) {
       return Message(event: PhoenixChannelEvent.error);
     }
     return null;
@@ -32,10 +33,20 @@ class PhoenixException implements Exception {
 
   @override
   String toString() {
-    if (socketError is PhoenixSocketErrorEvent) {
-      return socketError.error.toString();
+    if (socketError != null) {
+      return socketError!.error.toString();
     } else {
       return 'PhoenixException: socket closed';
     }
   }
+}
+
+/// Exception thrown when a [Push]'s reply was awaited for, and
+/// the allowed delay has passed.
+class ChannelTimeoutException implements Exception {
+  /// Default constructor
+  ChannelTimeoutException(this.response);
+
+  /// The PushReponse containing the timeout event.
+  PushResponse response;
 }