+++ lib/src/push.dart
@@ -1,16 +1,16 @@
 import 'dart:async';
 
+import 'package:equatable/equatable.dart';
 import 'package:logging/logging.dart';
 import 'package:quiver/collection.dart';
-import 'package:equatable/equatable.dart';
 
 import 'channel.dart';
 import 'events.dart';
-import 'exception.dart';
+import 'exceptions.dart';
 import 'message.dart';
 
 /// Encapsulates the response to a [Push].
-class PushResponse implements Equatable {
+class PushResponse extends Equatable {
   /// Builds a PushResponse from a status and response.
   PushResponse({
     this.status,
@@ -31,9 +31,9 @@ class PushResponse implements Equatable {
   /// }
   /// ```
   factory PushResponse.fromMessage(Message message) {
-    final data = message.payload;
+    final data = message.payload!;
     return PushResponse(
-      status: data['status'] as String,
+      status: data['status'] as String?,
       response: data['response'],
     );
   }
@@ -41,7 +41,7 @@ class PushResponse implements Equatable {
   /// Status provided by the backend.
   ///
   /// Value is usually either 'ok' or 'error'.
-  final String status;
+  final String? status;
 
   /// Arbitrary JSON content provided by the backend.
   final dynamic response;
@@ -56,12 +56,13 @@ class PushResponse implements Equatable {
   bool get isTimeout => status == 'timeout';
 
   @override
-  List<Object> get props => [status, response];
+  List<Object?> get props => [status, response];
 
   @override
   bool get stringify => true;
 }
 
+/// Type of function that should return a push payload
 typedef PayloadGetter = Map<String, dynamic> Function();
 
 /// Object produced by [PhoenixChannel.push] to encapsulate
@@ -83,31 +84,35 @@ class Push {
       ListMultimap();
 
   /// The event name associated with the pushed message
-  final PhoenixChannelEvent event;
+  final PhoenixChannelEvent? event;
 
   /// A getter function that yields the payload of the pushed message,
   /// usually a JSON object.
-  final PayloadGetter payload;
+  final PayloadGetter? payload;
 
   /// Channel through which the message was sent.
   final PhoenixChannel _channel;
 
   /// The expected timeout, after which the push is considered failed.
-  Duration timeout;
+  Duration? timeout;
 
-  PushResponse _received;
+  PushResponse? _received;
   bool _sent = false;
   bool _awaitingReply = false;
-  Timer _timeoutTimer;
-  String _ref;
-  PhoenixChannelEvent _replyEvent;
+  Timer? _timeoutTimer;
+  String? _ref;
+  PhoenixChannelEvent? _replyEvent;
 
-  Completer<PushResponse> _responseCompleter;
+  Completer<PushResponse>? _responseCompleter;
 
   /// A future that will yield the response to the original message.
-  Future<PushResponse> get future {
+  Future<PushResponse> get future async {
     _responseCompleter ??= Completer<PushResponse>();
-    return _responseCompleter.future;
+    final response = await _responseCompleter!.future;
+    if (response.isTimeout) {
+      throw ChannelTimeoutException(response);
+    }
+    return response;
   }
 
   /// Indicates whether the push has been sent.
@@ -133,7 +138,7 @@ class Push {
   ///
   /// This also schedules the timeout to be triggered in the future.
   Future<void> send() async {
-    if (_received is PushResponse && _received.isTimeout) {
+    if (_received is PushResponse && _received!.isTimeout) {
       _logger.warning('Trying to send push $ref after timeout');
       return;
     }
@@ -144,9 +149,9 @@ class Push {
     startTimeout();
     try {
       await _channel.socket.sendMessage(Message(
-        event: event,
+        event: event!,
         topic: _channel.topic,
-        payload: payload(),
+        payload: payload!(),
         ref: ref,
         joinRef: _channel.joinRef,
       ));
@@ -165,7 +170,7 @@ class Push {
   ///
   /// This is usually done automatically by the managing [PhoenixChannel]
   /// after a reconnection.
-  Future<void> resend(Duration newTimeout) async {
+  Future<void> resend(Duration? newTimeout) async {
     timeout = newTimeout ?? timeout;
     reset();
     await send();
@@ -185,12 +190,11 @@ class Push {
   void startTimeout() {
     if (!_awaitingReply) {
       _channel.onPushReply(replyEvent)
-        ..then(_receiveResponse)
-        ..catchError(_receiveResponse);
+        ..then<void>(_receiveResponse).catchError(_receiveResponse);
       _awaitingReply = true;
     }
 
-    _timeoutTimer ??= Timer(timeout, () {
+    _timeoutTimer ??= Timer(timeout!, () {
       _timeoutTimer = null;
       _logger.warning('Push $ref timed out');
       _channel.trigger(Message.timeoutFor(ref));
@@ -221,7 +225,7 @@ class Push {
     _received = response;
 
     if (_responseCompleter != null) {
-      if (_responseCompleter.isCompleted) {
+      if (_responseCompleter!.isCompleted) {
         _logger
           ..warning('Push being completed more than once')
           ..warning(
@@ -236,19 +240,22 @@ class Push {
         _logger.finer(
           () => 'Completing for $replyEvent with response ${response.response}',
         );
-        _responseCompleter.complete(response);
+        _responseCompleter!.complete(response);
       }
     }
+
     _logger.finer(() {
       if (_receivers[response.status].isNotEmpty) {
         return 'Triggering ${_receivers[response.status].length} callbacks';
       }
       return 'Not triggering any callbacks';
     });
-    for (final cb in _receivers[response.status]) {
+
+    final receivers = _receivers[response.status].toList();
+    clearWaiters();
+    for (final cb in receivers) {
       cb(response);
     }
-    clearWaiters();
   }
 
   /// Dispose the set of waiters and future associated with this push.
@@ -264,8 +271,8 @@ class Push {
         trigger(PushResponse.fromMessage(response));
       }
     } else if (response is PhoenixException) {
-      if (_responseCompleter is Completer && !_responseCompleter.isCompleted) {
-        _responseCompleter.completeError(response);
+      if (_responseCompleter is Completer && !_responseCompleter!.isCompleted) {
+        _responseCompleter!.completeError(response);
         clearWaiters();
       }
     }